Reimplemented and optimized counting by using dictionaries
Created a new class, DicePoolReporter in the file DiceDicts.py, that
successfully counts outcomes by using native dictionaries rather than
the naively implemented DieNode class and resulting tree of outcomes.

DicePoolReporter computes results which match the old DicePoolData class
(and also match 3rd party code and tools purporting to compute the same
probabilities), but it only presently computes the discrete results (and
not the cumulative results or a summary).

The old NarrativeDice.py file has not yet been removed so it can be
referenced and used to test the two methods against each other. It will
be removed once DicePoolReporter duplicates the full functionality of
DieNode and DicePoolData.